+++ packages/components/src/components/select/single-select/select.tsx
@@ -671,8 +671,6 @@ export class Choices implements IChoicesProps, IChoicesMethods {
       this.classList.remove('focus');
     });
 
-
-
   }